feat(Interpreter): a tree-walk Interpreter that evaluates simple expressions
The Interpreter is a class with an interpret() method, which accepts an expression
and evaluates it. To do so, each node in the AST is recursively traversed
post-order.

Runtime type-checking is performed for evaluating number operands; checks that fail
result in a RuntimeError exception that is reported to the main Lox class, which
will terminate the program if not running in shell mode.

Interesting design choices include implementing truthiness, equality comparisons
for the nil object representation.